src/requirements/common.txt
# Common dependencies for development and production
cryptography==2.3
django-common==0.1.51
django-extensions==1.7.7
django-extra-views==0.9.0
django-filter==1.0.4
django-split-settings==0.3.0
django==1.11.1
djangorestframework==3.6.3
docker==2.2.1
pika==0.10.0
pillow==4.1.1
psycopg2==2.7.1
pysolr==3.6.0
python-dateutil==2.6.1
python-json-logger==0.1.8
tblib==1.3.1
tzlocal==1.4
